+++ src/nupic/algorithms/Svm.hpp
@@ -178,7 +178,7 @@ public:
     std::copy(x_[i], x_[i] + n_dims(), sv);
   }
 
-  int persistent_size() const;
+  size_t persistent_size() const;
   void save(std::ostream &outStream) const;
   void load(std::istream &inStream);
 
@@ -202,7 +202,7 @@ private:
 
 //------------------------------------------------------------------------------
 struct svm_problem01
- {
+{
   typedef float label_type;
   typedef int feature_type;
 
@@ -233,11 +233,11 @@ struct svm_problem01
 
   inline ~svm_problem01() {
     if (recover_)
-      for (int i = 0; i != size(); ++i)
+      for (size_t i = 0; i != size(); ++i)
         delete[] x_[i];
   }
 
-  inline int size() const { return (int)x_.size(); }
+  inline size_t size() const { return x_.size(); }
   inline int n_dims() const { return n_dims_; }
   inline int nnz(int i) const { return nnz_[i]; }
 
@@ -260,7 +260,7 @@ struct svm_problem01
     while (x_it != x_end) {
       float val = *x_it;
       if (!nearlyZero(val, threshold_)) {
-        buf_[nnz] = x_it - x;
+        buf_[nnz] = (int)(x_it - x);
         ++nnz;
       }
       ++x_it;
@@ -290,14 +290,14 @@ struct svm_problem01
       sv[x_[i][k]] = 1;
   }
 
-  int persistent_size() const;
+  size_t persistent_size() const;
   void save(std::ostream &outStream) const;
   void load(std::istream &inStream);
 
   void print() const {
     std::cout << "Size = " << size() << " n dims = " << n_dims() << std::endl;
 
-    for (int i = 0; i != size(); ++i) {
+    for (size_t i = 0; i != size(); ++i) {
       std::cout << y_[i] << ": " << nnz_[i] << ": ";
       for (int j = 0; j != nnz_[i]; ++j)
         std::cout << x_[i][j] << " ";
@@ -346,7 +346,7 @@ public:
 
   ~svm_model();
 
-  int persistent_size() const;
+  size_t persistent_size() const;
   void save(std::ostream &outStream) const;
   void load(std::istream &inStream);
   void print() const;
@@ -669,7 +669,7 @@ private:
 
 public:
   QMatrix01(const svm_problem01 &prob, float g, int kernel, int cache_size)
-      : l(prob.size()), n(prob.n_dims()), kernel_function(nullptr), gamma(g),
+      : l((int)prob.size()), n(prob.n_dims()), kernel_function(nullptr), gamma(g),
         nnz(prob.nnz_), x(prob.x_.begin(), prob.x_.end()),
         x_square(new float[l]), y(new signed char[l]),
         cache(new Cache<float>(l, (long int)(cache_size * (1 << 20)))),
@@ -758,7 +758,7 @@ struct svm_parameter {
   std::vector<int> weight_label;
   std::vector<float> weight;
 
-  int persistent_size() const;
+  size_t persistent_size() const;
   void save(std::ostream &outStream) const;
   void load(std::istream &inStream);
 
@@ -871,7 +871,7 @@ public:
 
   float cross_validation(int);
 
-  int persistent_size() const;
+  size_t persistent_size() const;
   void save(std::ostream &outStream) const;
   void load(std::istream &inStream);
 
@@ -947,7 +947,7 @@ public:
     return svm_.cross_validation(n_fold);
   }
 
-  inline int persistent_size() const { return svm_.persistent_size(); }
+  inline size_t persistent_size() const { return svm_.persistent_size(); }
 
   inline void save(std::ostream &outStream) const { svm_.save(outStream); }
